// src/app/acnhapi/models/name.interface.ts
export interface NameInterface {
  nameUSen: string;
  nameEUen: string;
  nameEUde: string;
  nameEUes: string;
  nameUSes: string;
  nameEUfr: string;
  nameUSfr: string;
  nameEUit: string;
  nameEUnl: string;
  nameCNzh: string;
  nameTWzh: string;
  nameJPja: string;
  nameKRko: string;
  nameEUru: string;
}